% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAssociationTest.R
\name{jat}
\alias{jat}
\title{Jointly testing for main and interaction effects with splicing of the TE of one gene on the TE of another}
\usage{
jat(y, x, Z, IER, ...)
}
\arguments{
\item{y}{gene expression level of the response gene Y. It is a standardized numeric vector.}

\item{x}{gene expression level of the regulator gene X. It is a standardized numeric vector.}

\item{Z}{covariates needed to be adjusted. It is a numeric vector or matrix.}

\item{IER}{intron excision ratio matrix of regulator gene X obtained from leafcutter or \link[ASTest]{JuncToIER}.}

\item{...}{additional arguments.}
}
\value{
A list containing:
\item{CALL}{function call.}
\item{data}{data used.}
\item{formula}{model under null hypothesis.}
\item{P}{the p-value of the joint effects of both fixed and random effects of the TE of one gene on the TE of another.}
\item{Q}{a score test statistic, which follows a mixed chi-square distribution.}
\item{rho}{relative contribution of the fixed effects and random effects in the joint association test.}
\item{residual}{residuals of the model under null hypothesis.}
\item{sigmau}{estimated variance component of the random intercept.}
\item{sigmae}{estimated variance component of the error term.}
\item{sigmaint}{estimated variance component of the random slope of TE.}
}
\description{
\code{jat} estimates the variance components and tests the splicing-TE interaction, as well as the fixed effect of TE jointly with a variance-component score test.
}
\details{
This function \code{jat} jointly tested the fixed effect of TE and interaction effect of TE and splicing on other genes.
It was implemented based on the StructLMM method and used the GRAMMAR-GAMMA functions to estimate variance components (see reference).
Here we adopted a mixed-effects model in estimating the co-expression effects between pairs of genes while allowing sample-level variation as random effects,
with the covariance matrix of the random effects being proportional to sample-sample covariance of isoform composition.
}
\references{
Moore, R., Casale, F.P., Jan Bonder, M. et al. A linear mixed-model approach to study multivariate gene–environment interactions. Nat Genet 51, 180–186 (2019). \url{https://doi.org/10.1038/s41588-018-0271-0}

Svishcheva, G., Axenovich, T., Belonogova, N. et al. Rapid variance components–based method for whole-genome association analysis. Nat Genet 44, 1166–1170 (2012). \url{https://doi.org/10.1038/ng.2410}
}
