% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhoEstimation.R
\name{RhoEstimation}
\alias{RhoEstimation}
\title{Estimation of relative contribution of the fixed effects and random effects.}
\usage{
RhoEstimation(y, x, Z, IER)
}
\arguments{
\item{y}{gene expression level of the response gene Y. It is a standardized numeric vector.}

\item{x}{gene expression level of the regulator gene X. It is a standardized numeric vector.}

\item{Z}{covariates needed to be adjusted. It is a numeric vector or matrix.}

\item{IER}{intron excision ratio matrix of regulator gene X obtained from leafcutter or \link[ASTest]{JuncToIER}.}

\item{...}{additional arguments.}
}
\value{
A list containing:
\item{rho}{relative contribution of the fixed effects and random effects in the joint association test.}
\item{sigmaint}{estimated variance component of the random slope of TE.}
}
\description{
\code{RhoEstimation} estimates the relative contribution of the fixed effects and random effects. It returns \eqn{\rho} which is informative in characterizing the relative contributions of main (fixed) effects and isoform-specific TE- TE effects (interaction/random effects) when studying regulation patterns of co-expressed genes.
}
\details{
When \eqn{\rho = 0}, the joint association test is equivalent to the standard correlation test adjusting for IER and there is no splicing-induced sample-level variation in TE-TE effect. When ρ is large, the splicing-induced random slope heavily affects the variation of TE of gene Y.
}
